Drop the legacy core-backend and registry impls from backend-postgres
