+++ Cargo.toml
@@ -154,7 +154,6 @@ waymark-workflow-completion-convert-proto = { path = "crates/lib/workflow-comple
 waymark-workflow-completion-core = { path = "crates/lib/workflow-completion-core" }
 waymark-workflow-completion-direct = { path = "crates/lib/workflow-completion-direct" }
 waymark-workflow-initialization-convert-proto = { path = "crates/lib/workflow-initialization-convert-proto" }
-waymark-workflow-registry-backend = { path = "crates/lib/workflow-registry-backend" }
 waymark-workflow-service-vm-executables = { path = "crates/lib/workflow-service-vm-executables" }
 waymark-workflow-service-vm-executables-backend = { path = "crates/lib/workflow-service-vm-executables-backend" }
 waymark-workflow-service-vm-runtimes = { path = "crates/lib/workflow-service-vm-runtimes" }
