--- crates/lib/workflow-registry-backend/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "waymark-workflow-registry-backend"
-version.workspace = true
-publish.workspace = true
-edition = "2024"
-
-[features]
-default = ["either"]
-
-[dependencies]
-waymark-backends-core = { workspace = true }
-waymark-ids = { workspace = true }
-
-either = { workspace = true, optional = true }
-
-[lib]
-test = false
-doctest = false
